Take another stab at the AI player.  Largely fail, but learn a few things, maybe, and come up with some more ideas for later.
I think that a more heuristically approach combined with minimax is my best bet.  I can't program all the logic that goes through my own head why I play the game.  I need a way to intuitively guess where to concentrate minimax.  And minimax's evaluation function needs to be improved, and we need to push it to the limit.  Currently, minimax looks at liberties only, but it should also look at territory and captures, and all of these things for both players.